requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
python-dateutil==2.9.0.post0
pytest==8.4.1
